fix autocomplete after dot when next line is a comment
now simply insert a `;` after a `.` if it's the last character in line.
this could break syntax when this is done on purpose (define the property on next line), but why would you do that?

having a `;` makes more sense anyway:
```coffeescript
a.|
console.log 123
```
only used to yield proper syntax completion at `|` because this snippet was falsely translated into
```js
a.console.log(123);
```
which obviously does not make sense (although it worked for its purposes). And it's also why a autocompletion with comment in the next line failed.
Now, with the semicolon, the error is provoked to occur at the dot line itself, triggering fakeLine logic in transpileService, succeeding compilation and giving functioning autocomplete regardless of the content of the following lines.