don't postprocess var-assignment-lines that have been added by the CS compiler, but only var-declarations
these assignments (`var a = ...;`) are a new feature of the recently switched coffeescript compiler and were falsely being read as declarations (`var a, b;`). This did not really change the code, as the respective implementations can never exist and be altered, but it's firstly unnecessary computation and 2. the indent was removed, resulting in bad source maps.
these bad source maps/indentations are still there but it doesn't matter, as declaration lines column mapping must never be relevant to coffeesense.